# Conflicts:
#	code/Parknshop/src/main/webapp/WEB-INF/web.xml

+++ code/Parknshop/src/main/java/com/tmazon/service/impl/OrderServiceImpl.java
@@ -1,17 +1,16 @@
 package com.tmazon.service.impl;
 
-import java.util.Date;
+import java.sql.Timestamp;
 import java.util.List;
 
 import com.tmazon.dao.CartDao;
+import com.tmazon.dao.OrderDao;
 import com.tmazon.dao.OrderInfoDao;
 import com.tmazon.dao.ProductDao;
-import com.tmazon.dao.ShopDao;
-import com.tmazon.dao.OrderDao;
 import com.tmazon.domain.Cart;
+import com.tmazon.domain.Order;
 import com.tmazon.domain.OrderInfo;
 import com.tmazon.domain.Product;
-import com.tmazon.domain.Order;
 import com.tmazon.domain.User;
 import com.tmazon.service.OrderService;
 import com.tmazon.util.BasicFactory;
@@ -35,7 +34,7 @@ public class OrderServiceImpl implements OrderService {
 
 		boolean flag = true;
 
-		order.setOrderTime(new Date());
+		order.setOrderTime(new Timestamp(System.currentTimeMillis()));
 		order.setStatus(Order.STATUS_NORMAL);
 		order.setPayType(Order.PAY_TYPE_ONLINE_PAYMENT);
 
@@ -46,21 +45,24 @@ public class OrderServiceImpl implements OrderService {
 			Cart cart = cartDao.findByPKId(order.getUserId(),
 					orderInfo.getProductId());
 			Product product = productDao.findById(orderInfo.getProductId());
-			int stockNum = product.getSoldNum();
+			int stockNum = product.getStockNum();
 			int quantity = cart.getQuantity();
 			if(stockNum < quantity){
 				cart.setQuantity(stockNum);
 			}
 			orderInfo.setQuantity(cart.getQuantity());
 			orderInfo.setStatus(OrderInfo.STATUS_UNPAID);
-			if (orderInfoDao.insert(orderInfo)) {
-				product.setStockNum(product.getStockNum() - cart.getQuantity());
-				product.setSoldNum(product.getSoldNum() + cart.getQuantity());
-				productDao.updateStockNum(product);
-				cartDao.delete(new Cart(order.getUserId(), orderInfo
-						.getProductId(), null));
-			} else {
-				flag = false;
+			if (orderInfo.getQuantity() != 0) {
+				if(orderInfoDao.insert(orderInfo)){
+					int soldNum = product.getSoldNum() + 0;
+					product.setStockNum(product.getStockNum() - cart.getQuantity());
+					product.setSoldNum(soldNum + cart.getQuantity());
+					productDao.updateStockNum(product);
+					cartDao.delete(new Cart(order.getUserId(), orderInfo
+							.getProductId(), null));
+				}else {
+					flag = false;
+				}
 			}
 		}
